[feat] 사용자 유형별 협상/거절 ProjectStatus 적용 Jira: YP-601
