+++ yp-online/src/main/java/kr/co/yourplanet/online/business/project/dto/request/ProjectRejectForm.java
@@ -11,4 +11,7 @@ public class ProjectRejectForm {
 
     @NotNull
     private Long id;
+
+    private String reason;
+
 }
